% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{bellmanford}
\alias{bellmanford}
\title{Bellman-Ford algorithm (R version)}
\usage{
bellmanford(summits, edges, start)
}
\arguments{
\item{summits}{a vector of all summits}

\item{edges}{a nested list, which items are composed of `from`, `to` and `weight`}

\item{start}{the node which distance from is computed}
}
\value{
a list with `d` the constructed distance for each node and `pred` a vector containing at item i the predecessor of i-th summit
}
\description{
Find shortest path between two nodes in a graph
}
